Problem with the counter.
Hello Nick,
I manage to use the counter to modulate the sound as I want thanks to the even and odd number of revolutions;)
But when I change the direction of rotation with the 'o' key and go to position 0, the counter does not want to increase or decrease the rotation of the numbers. It's a shame.

I hope I sent the program well this time. (Should I write the program always with the same name?)